Controller has been decoupled from model and view.
View is now an interface with implementable methods. The concrete view is injected by the control class. The view bean has listeners which use the commands which in turn are coupled to the Facade, thus the view is auto injected to Facade due to the recursive injections starting with the concrete view. All input  validation takes place in the Facade now. The Facade is more the control class than the control class. This supports my theory for platform independence.
Signed-off-by:Muhammad Callender <dreadmorpheus@gmail.com>